Refactor widgets and classes out of map_page.dart and panel_page.dart and into subdirectories Rename hazard_info.dart to hazard_info_card.dart Make all imports absolute rather than relative
