+++ lib/consts.dart
@@ -1,4 +1,3 @@
-import 'package:flutter/cupertino.dart';
 import 'package:flutter/foundation.dart';
 import 'package:forest_park_reports/model/camera_position.dart';
 import 'package:intl/intl.dart';
