+++ lib/page/home_page/panel_page/hazard_info_card.dart
@@ -1,6 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:forest_park_reports/model/hazard_update.dart';
-import 'package:forest_park_reports/page/home_page/map_page.dart';
+
+import 'hazard_image.dart';
 
 class UpdateInfoWidget extends StatelessWidget {
   final HazardUpdateModel update;
